+++ src/DataGrid.tsx
@@ -118,7 +118,7 @@ export interface DataGridProps<R, K extends keyof R> {
    * Toggles and modes
    */
   /** Toggles whether filters row is displayed or not */
-  enableHeaderFilters?: boolean;
+  enableFilters?: boolean;
   /** Toggles whether cells should be autofocused */
   enableCellAutoFocus?: boolean;
   enableCellCopyPaste?: boolean;
@@ -150,7 +150,7 @@ function DataGrid<R, K extends keyof R>({
   height = 350,
   width,
   enableCellAutoFocus = true,
-  enableHeaderFilters = false,
+  enableFilters = false,
   enableCellCopyPaste = false,
   enableCellDragAndDrop = false,
   cellNavigationMode = CellNavigationMode.NONE,
@@ -240,7 +240,7 @@ function DataGrid<R, K extends keyof R>({
     props.onRowsUpdate?.(event);
   }
 
-  const rowOffsetHeight = headerRowHeight + (enableHeaderFilters ? headerFiltersHeight : 0);
+  const rowOffsetHeight = headerRowHeight + (enableFilters ? headerFiltersHeight : 0);
 
   return (
     <div
@@ -271,7 +271,7 @@ function DataGrid<R, K extends keyof R>({
               onSort={props.onSort}
               scrollLeft={nonStickyScrollLeft}
             />
-            {enableHeaderFilters && (
+            {enableFilters && (
               <FilterRow<R, K>
                 height={headerFiltersHeight}
                 width={columnMetrics.totalColumnWidth + getScrollbarSize()}